+++ Services/MyFitScope.Services.Data/UsersService.cs
@@ -1,5 +1,6 @@
 ﻿namespace MyFitScope.Services.Data
 {
+    using System;
     using System.Threading.Tasks;
 
     using Microsoft.AspNetCore.Http;
@@ -9,6 +10,9 @@
 
     public class UsersService : IUsersService
     {
+        private const string InvalidUserIdErrorMessage = "User with ID: {0} does not exist.";
+        private const string InvalidCloudinaryResponseParams = "Missing response from Cloudinary!";
+
         private readonly ICloudinaryService cloudinaryService;
         private readonly IDeletableEntityRepository<ApplicationUser> usersRepository;
 
@@ -22,12 +26,23 @@
         {
             var user = await this.usersRepository.GetByIdWithDeletedAsync(loggedInUserId);
 
+            if (user == null)
+            {
+                throw new ArgumentNullException(
+                    string.Format(InvalidUserIdErrorMessage, loggedInUserId));
+            }
+
             if (user.AvatarImageUrl != null)
             {
                 this.cloudinaryService.DeletePhoto(user.AvatarImagePublicId);
             }
 
             var uploadPhotoResponse = await this.cloudinaryService.UploadPhotoAsync(file, user.UserName, GlobalConstants.CloudUsersImageFolder);
+
+            if (uploadPhotoResponse == null)
+            {
+                throw new ArgumentNullException(InvalidCloudinaryResponseParams);
+            }
 
             user.AvatarImagePublicId = uploadPhotoResponse.PublicId;
             user.AvatarImageUrl = uploadPhotoResponse.PhotoUrl;
